changed evaluation value to be long

+++ ChessEngineVirtualPlayer/src/main/java/com/davidbalazs/chess/algorithms/EvaluationFunction.java
@@ -6,5 +6,5 @@ import com.davidbalazs.chess.model.ChessPosition;
  * @author: david.balazs@iquestgroup.com
  */
 public interface EvaluationFunction {
-    int evaluate(ChessPosition chessPosition);
+    long evaluate(ChessPosition chessPosition);
 }
